+++ api/src/opentrons/protocol_engine/commands/__init__.py
@@ -145,6 +145,21 @@ from .set_rail_lights import (
     SetRailLightsCommandType,
 )
 
+from .touch_tip import (
+    TouchTip,
+    TouchTipParams,
+    TouchTipCreate,
+    TouchTipResult,
+    TouchTipCommandType,
+)
+
+from .blow_out import (
+    BlowOutParams,
+    BlowOutResult,
+    BlowOutCreate,
+    BlowOutImplementation,
+    BlowOut,
+)
 
 __all__ = [
     # command type unions
@@ -242,6 +257,18 @@ __all__ = [
     "SetRailLightsCreate",
     "SetRailLightsResult",
     "SetRailLightsCommandType",
+    # touch tip command models
+    "TouchTip",
+    "TouchTipParams",
+    "TouchTipCreate",
+    "TouchTipResult",
+    "TouchTipCommandType",
+    # blow out command models
+    "BlowOutResult",
+    "BlowOutCreate",
+    "BlowOutImplementation",
+    "BlowOutParams",
+    "BlowOut",
     # module command bundles
     "heater_shaker",
     "magnetic_module",