+++ api/src/opentrons/protocol_runner/json_command_translator.py
@@ -107,8 +107,6 @@ class JsonCommandTranslator:
         exclude_commands = [
             "loadLiquid",
             "delay",
-            "touchTip",
-            "blowout",
             "moveToSlot",
             "moveToCoordinates",
         ]