+++ api/tests/opentrons/protocol_runner/test_json_command_translator.py
@@ -28,135 +28,6 @@ from opentrons.protocol_engine import (
     ModuleLocation,
 )
 
-INVALID_TEST_PARAMS = [
-    (
-        protocol_schema_v6.Command(
-            commandType="aspirate",
-            params=protocol_schema_v6.Params(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                volume=1.23,
-                # todo (Max and Tamar 3/17/22):needs to be added to the aspirate command
-                # https://github.com/Opentrons/opentrons/issues/8204
-                flowRate=4.56,
-                wellName="A1",
-                wellLocation=protocol_schema_v6.WellLocation(
-                    origin="bottom",
-                    offset=protocol_schema_v6.OffsetVector(x=0, y=0, z=7.89),
-                ),
-            ),
-        ),
-        pe_commands.AspirateCreate(
-            params=pe_commands.AspirateParams(
-                # todo: id
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                volume=1.23,
-                wellName="A1",
-                wellLocation=WellLocation(),
-            )
-        ),
-    ),
-    (
-        protocol_schema_v6.Command(
-            commandType="dispense",
-            params=protocol_schema_v6.Params(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                volume=1.23,
-                flowRate=4.56,
-                wellName="A1",
-                wellLocation=protocol_schema_v6.WellLocation(
-                    origin="bottom",
-                    offset=protocol_schema_v6.OffsetVector(x=0, y=0, z=7.89),
-                ),
-            ),
-        ),
-        pe_commands.DispenseCreate(
-            params=pe_commands.DispenseParams(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                volume=1.23,
-                wellName="A1",
-                wellLocation=WellLocation(),
-            )
-        ),
-    ),
-    (
-        protocol_schema_v6.Command(
-            commandType="dropTip",
-            params=protocol_schema_v6.Params(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                wellName="A1",
-                wellLocation=protocol_schema_v6.WellLocation(
-                    origin="bottom",
-                    offset=protocol_schema_v6.OffsetVector(x=0, y=0, z=7.89),
-                ),
-            ),
-        ),
-        pe_commands.DropTipCreate(
-            params=pe_commands.DropTipParams(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                wellName="A1",
-                wellLocation=WellLocation(),
-            )
-        ),
-    ),
-    (
-        protocol_schema_v6.Command(
-            commandType="pickUpTip",
-            params=protocol_schema_v6.Params(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                wellName="A1",
-                wellLocation=protocol_schema_v6.WellLocation(
-                    origin="bottom",
-                    offset=protocol_schema_v6.OffsetVector(x=0, y=0, z=7.89),
-                ),
-            ),
-        ),
-        pe_commands.PickUpTipCreate(
-            params=pe_commands.PickUpTipParams(
-                pipetteId="pipette-id-abc123",
-                labwareId="labware-id-def456",
-                wellName="A1",
-                wellLocation=WellLocation(),
-            )
-        ),
-    ),
-    (
-        protocol_schema_v6.Command(
-            commandType="loadPipette",
-            params=protocol_schema_v6.Params(pipetteId="pipetteId", mount="left"),
-        ),
-        pe_commands.LoadPipetteCreate(
-            params=pe_commands.LoadPipetteParams(
-                pipetteId="pipetteId",
-                pipetteName=PipetteName("p10_single"),
-                mount=MountType("right"),
-            )
-        ),
-    ),
-    (
-        protocol_schema_v6.Command(
-            commandType="loadModule",
-            params=protocol_schema_v6.Params(
-                moduleId="magneticModuleId",
-                location=protocol_schema_v6.Location(slotName="3"),
-            ),
-        ),
-        pe_commands.LoadModuleCreate(
-            params=pe_commands.LoadModuleParams(
-                model=ModuleModel("magneticModuleV2"),
-                moduleId="magneticModuleId",
-                location=DeckSlotLocation(slotName=(DeckSlotName("4"))),
-            )
-        ),
-    ),
-]
-
 VALID_TEST_PARAMS = [
     (
         protocol_schema_v6.Command(
@@ -255,6 +126,31 @@ VALID_TEST_PARAMS = [
     ),
     (
         protocol_schema_v6.Command(
+            commandType="touchTip",
+            params=protocol_schema_v6.Params(
+                pipetteId="pipette-id-abc123",
+                labwareId="labware-id-def456",
+                wellName="A1",
+                wellLocation=protocol_schema_v6.WellLocation(
+                    origin="bottom",
+                    offset=protocol_schema_v6.OffsetVector(x=0, y=0, z=-1.23),
+                ),
+            ),
+        ),
+        pe_commands.TouchTipCreate(
+            params=pe_commands.TouchTipParams(
+                pipetteId="pipette-id-abc123",
+                labwareId="labware-id-def456",
+                wellName="A1",
+                wellLocation=WellLocation(
+                    origin=WellOrigin.BOTTOM,
+                    offset=WellOffset(x=0, y=0, z=-1.23),
+                ),
+            )
+        ),
+    ),
+    (
+        protocol_schema_v6.Command(
             commandType="pause",
             params=protocol_schema_v6.Params(
                 wait=True,
@@ -308,6 +204,31 @@ VALID_TEST_PARAMS = [
                 location=ModuleLocation(moduleId="temperatureModuleId"),
                 version=1,
                 namespace="example",
+            )
+        ),
+    ),
+    (
+        protocol_schema_v6.Command(
+            commandType="blowout",
+            params=protocol_schema_v6.Params(
+                pipetteId="pipette-id-abc123",
+                labwareId="labware-id-def456",
+                wellName="A1",
+                wellLocation=protocol_schema_v6.WellLocation(
+                    origin="bottom",
+                    offset=protocol_schema_v6.OffsetVector(x=0, y=0, z=7.89),
+                ),
+            ),
+        ),
+        pe_commands.BlowOutCreate(
+            params=pe_commands.BlowOutParams(
+                pipetteId="pipette-id-abc123",
+                labwareId="labware-id-def456",
+                wellName="A1",
+                wellLocation=WellLocation(
+                    origin=WellOrigin.BOTTOM,
+                    offset=WellOffset(x=0, y=0, z=7.89),
+                ),
             )
         ),
     ),
@@ -397,16 +318,3 @@ def test_load_command(
     """Test translating v6 commands to protocol engine commands."""
     output = subject.translate(_make_json_protocol(commands=[test_input]))
     assert output == [expected_output]
-
-
-@pytest.mark.parametrize("test_input, expected_output", INVALID_TEST_PARAMS)
-def test_invalid_commands(
-    subject: JsonCommandTranslator,
-    test_input: protocol_schema_v6.Command,
-    expected_output: pe_commands.CommandCreate,
-) -> None:
-    """Test fail invalid payload-translating v6 commands to protocol engine commands."""
-    with pytest.raises(AssertionError):
-        assert subject.translate(_make_json_protocol(commands=[test_input])) == [
-            expected_output
-        ]